+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **New Features:**
 - Added full language alias support (allows aliases names to be added that point to a specific language name).
+- Added full custom buttons support via a new attribute called "data-syntax-buttons".
 
 #### **Languages:**
 - Added support for the F# programming language (importable file).